fix: api 인스턴스 파일 이동
